Calculates PK parameters that need lambda_z

Calculates PK parameters that need lambda_z

calc.par.th(x = par, th = th, cov = cov, dosevar = "dose",
  factor = 1, reg = "sd", ss = "n")

Arguments

x
result parameter dataset from calc.par

th
result dataset from est.thalf

cov
covariates dataset (containing at least dose for CL calculation)

dosevar
variable containing the dose amount

factor
conversion factor for CL and V calculation (e.g. dose in mg, conc in ng/mL, factor=1000)

reg
regimen, "sd" or "md"

ss
is steady state reached (y/n)

Value

dataset with estimates for the following parameters, one observation per subject:
all parameters calculated in th
all parameters calculated in par
clast.pred: predicted concentration at tlast
aucinf.obs: aucinf based on observed concentration at tlast
aucinf.pred: aucinf based on predicted concentration at tlast
aumcinf.obs: area under the first moment curve extrapolated to infinity, based on observed concentration at tlast
aumcinf.pred: area under the first moment curve extrapolated to infinity, based on predicted concentration at tlast
cl.f.obs: clearance based on aucinf.obs, at steady state based on auctau
cl.f.pred: clearance based on aucinf.pred
mrt.obs: Mean residence time based on aumcinf.obs and aucinf.obs
mrt.pred: Mean residence time based on aumcinf.pred and aucinf.pred
vz.f.obs: distribution volume based on cl.f.obs, at steady state based on auctau
vz.f.pred: distribution based on cl.f.pred
vss.obs: Steady-state volume based on cl.obs and mrt.obs
vss.pred: Steady-state volume based on cl.pred and mrt.pred
regimen (reg)
steady state reached Y/N? (ss)
NOTE: ctmax must be merged separately as those were calculated from uncorrected data

Examples

library(dplyr)
#'# We need half-lives for corrections, so first let's get that.
th = Theoph %>%
 group_by(Subject=as.numeric(Subject)) %>%
 do(est.thalf(.,timevar="Time",depvar="conc",includeCmax="Y")) %>%
 ungroup()

# We need nominal time variable as well, so let's generate that.
ID <- as.numeric(Theoph$Subject)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph %>% mutate(NTAD=metrumrg::snap(Time, NTAD))

#let's say we want AUC0-8. We only have 7 and 9 hr concentrations, so we need to interpolate conc for 8 hr.
tc = Theoph1 %>%
group_by(Subject=as.numeric(Subject)) %>%
do(correct.time(.,nomtimevar="NTAD",timevar="Time",depvar="conc",
                 tau=,tstart=,tend=,teval=8,th=th,reg="sd")) %>%
#above step added timepoints that we will add interpolated concentrations to with this next step
do(correct.conc(.,nomtimevar="NTAD",tau=,tstart=,tend=,teval=8,
th=th,reg="sd",ss="n")) %>%
ungroup()

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#Now get parameters that do not require LAMBDAZ
par <- tc %>%
     group_by(Subject=as.numeric(Subject)) %>%
     ##tstart by default is 0, so here we will get AUC0-8hr and AUC0-24 as sampling ends at 24 hr
     do(calc.par(.,tau=NA, tstart=NA, tend=NA, teval=8, route="po", method=1)) %>%
     ungroup()
cov <- data.frame(Subject=as.numeric(Theoph1$Subject), DOSE=Theoph$Dose) %>%
       distinct(.,.keep_all = T)
par <- calc.par.th(x=par,th=th ,cov=cov,
                 dose="DOSE",factor=1,
                 reg="sd",ss="n")

#> Joining, by = "Subject"

#> Joining, by = "Subject"