Corrects missing concentration at critical time points (e.g, predose, TAU, start and end of user selected AUC interval)

If there is a measurable concentration BEFORE and AFTER the missing concentration, use interpolation If there is NO measurable concentration AFTER the missing concentration, use extrapolation Set missing concentration at predose to 0 (SD, non-endogenous) or value at t=TAU (steady state only) Set missing concentration at t=TAU to value at t=0 (steady state only) The following Concentration Deviation Correction Rules will be applied to critical time points (t=0, tau, tstart, tend, teval), if needed:

Rule	Regimen	Description

Applied to		SDC-1	sd

Set concentration to 0 (only non-endogenous compounds)	t=0		SDC-2

sd	impute missing concentration by interpolation	t=tau,tstart,tend,teval

SDC-3	sd	impute missing concentration by extrapolation	t=tau,tend,teval

SDC-4	sd (IV)	impute missing concentration by back-extrapolation

t=0		MDC-1	md

impute missing concentration by existing conc at t=0 or t=tau*	t=0,tau		MDC-2

md	impute missing concentration by interpolation	t=tau,tstart,tend,teval

MDC-3	md	impute missing concentration by extrapolation	t=tau,tend,teval

MDC-4	md (IV)	impute missing concentration by back-extrapolation

t=0		* only if steady state has been reached

correct.conc(x, nomtimevar = "ntad", tau = NA, tstart = NA,
  tend = NA, teval = NA, th = NA, reg = "sd", ss = "n",
  route = "po", method = 1)

Arguments

x
input dataset name input dataset name (contains all data, including LOQ (set conc to zero for these))

nomtimevar
variable name containing the nominal sampling time

tau
dosing interval (for multiple dosing), if single dose, leave empty

tstart
start time of partial AUC (start>0), if not requested, leave empty

tend
end time of partial AUC, if not requested, leave empty

teval
user selected AUC interval, if not requested, leave empty

th
file name of file with lamdba_z information for each curve (can be derived from est.thalf)

reg
regimen, "sd" or "md"

ss
is steady state reached (y/n)

route
route of drug administration ("po","iv")

method
of interpolation:
1: linear up - linear down
2: linear up - logarithmic down

Value

a dataset with missing concentrations imputed. The following variables are added:

crule.nr

correction rule number

crule.txt	text explaining what was altered

applies.to.conc

lists all AUCS to which the concentration correction rule applies

Examples

library(dplyr)
# We need half-lives for this, so first let's get that.
th = Theoph %>%
 group_by(Subject=as.numeric(Subject)) %>%
 do(est.thalf(.,timevar="Time",depvar="conc",includeCmax="Y")) %>%
 ungroup()

# We need nominal time variable as well, so let's generate that.
ID <- as.numeric(Theoph$Subject)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph %>% mutate(NTAD=metrumrg::snap(Time, NTAD))

#let's say we want AUC0-8. We only have 7 and 9 hr concentrations, so we need to interpolate conc for 8 hr.
tc = Theoph1 %>%
group_by(Subject=as.numeric(Subject)) %>%
do(correct.time(.,nomtimevar="NTAD",timevar="Time",depvar="conc",
                 tau=,tstart=,tend=,teval=8,th=th,reg="sd")) %>%
#above step added timepoints that we will add interpolated concentrations to with this next step
do(correct.conc(.,nomtimevar="NTAD",tau=,tstart=,tend=,teval=8,
th=th,reg="sd",ss="n")) %>%
ungroup()

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

#> Joining, by = c("Subject", "lambda_z", "start_th", "end_th", "includeCmax")

head(tc)

#> # A tibble: 6 x 56
#>   Subject    Wt  Dose  Time  conc  NTAD lambda_z start_th end_th includeCmax
#>     <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>  <dbl> <chr>      
#> 1       1    80     4 0      0      0     0.0916     9.22   23.8 Y          
#> 2       1    80     4 0.27   1.29   0.3   0.0916     9.22   23.8 Y          
#> 3       1    80     4 0.580  3.08   0.5   0.0916     9.22   23.8 Y          
#> 4       1    80     4 1.15   6.44   1     0.0916     9.22   23.8 Y          
#> 5       1    80     4 2.03   6.32   2     0.0916     9.22   23.8 Y          
#> 6       1    80     4 3.57   5.53   4     0.0916     9.22   23.8 Y          
#> # ... with 46 more variables: create.nr <chr>, create.txt <chr>,
#> #   trule.nr <chr>, trule.txt <chr>, applies.to.time <chr>, t0.flag <dbl>,
#> #   tau.flag <dbl>, tstart.flag <dbl>, tend.flag <dbl>, teval.flag <dbl>,
#> #   missflag <dbl>, misstime <dbl>, diff <dbl>, flag <dbl>, leaddv <dbl>,
#> #   lagdv <dbl>, leadtime <dbl>, lagtime <dbl>, conc.tau <dbl>, time.tau <dbl>,
#> #   conc.teval <dbl>, time.teval <dbl>, conc.part <dbl>, time.part <dbl>,
#> #   conc.lastall <dbl>, time.lastall <dbl>, newdepvar <dbl>, ptime <dbl>,
#> #   crule.nr <chr>, crule.txt <chr>, applies.to.conc <chr>, depvar <dbl>,
#> #   timevar <dbl>, lead.cteval <dbl>, lag.cteval <dbl>, lead.tteval <dbl>,
#> #   lag.tteval <dbl>, t0val <dbl>, tauval <lgl>, back_extrap <dbl>, lc1 <dbl>,
#> #   lc2 <dbl>, lt1 <dbl>, lt2 <dbl>, firstmeasc <dbl>, firstmeast <dbl>