Corrects concentrations at critical, but deviating time points (e.g, predose, TAU, start and end of user selected AUC interval), and adds missing records at these critical time points.

Records with missing NOMINAL time will be removed, this must be corrected before the function is called. If a record at the critical time point is missing, add it and set time to nominal time and set dv conc to NA If there is a measurable concentration AFTER the nominal time point (i.e. sample is taken too late), use interpolation If there is NO measurable concentration AFTER the nominal time point (i.e. sample is taken too early), use extrapolation Set deviating time at predose to 0 Original time and conc will be kept in original variables. The following Time Deviation Correction Rules will be applied to critical time points (t=0, tau, tstart, tend, teval), if needed:

Rule	Regimen	Description	Applied to

SDT-1	sd	Set actual time to 0	t=0

SDT-2	sd	Correct concentration at deviating time by interpolation	t=tau,tstart,tend,teval

SDT-3	sd	Correct concentration at deviating time by extrapolation	t=tau,tend,teval

MDT-1	md	If predose sample taken after dosing, set actual time to 0 and conc to NA	t=0

MDT-2	md	Correct concentration at deviating time by interpolation (too late)	t=tau,tstart,tend,teval

MDT-3	md	Correct concentration at deviating time by extrapolation (too early)	t=0,tau,tend,teval

MDT-3a	md	Set actual time to zero if concentration is BLOQ (too early)	t=0

correct.time(x, nomtimevar = "ntad", timevar = "time", depvar = "dv",
  tau = NA, tstart = NA, tend = NA, teval = NA, th = NA,
  reg = "sd", method = 1)

Arguments

x
input dataset name (contains all data, including LOQ (set conc to zero for these))

nomtimevar
variable name containing the nominal sampling time

timevar
variable name containing the actual sampling time

depvar
variable name containing the dependent variable (e.g., concentration)

tau
dosing interval (for multiple dosing), if single dose, leave empty

teval
user selected AUC interval, if not requested, leave empty

th
file name of file with lamdba_z information for each curve (can be derived from est.thalf)

reg
regimen, "sd" or "md"

method
method of interpolation:
1: linear up - linear down
2: linear up - logarithmic down

Value

a dataset with time deviation corrections applied (timevar and depvar adapted). The following variables are added:

create.nr	is a missing record created?

create.txt	explanation of what is created

trule.nr	correction rule number

trule.txt	text explaining what was altered

applies.to.time	lists all AUCS to which the time deviation rule applies

time.tau, conc.tau	time and conc, corrected for AUCtau calculation

time.teval, conc.teval	time and conc, corrected for AUCteval calculation (AUC0-teval)

time.part, conc.part	time and conc, corrected for partial AUC calculation (AUCstart-end, start>0)

time.lastall, conc.lastall	time and conc, corrected for AUClast and AUCall calculation

t0.flag, tau.flag, tstart.flag, tend.flag, teval.flag	flags for what timepoint the correction was needed

Examples

library(dplyr)
# We need half-lives for this, so first let's get that.
th = Theoph %>%
group_by(Subject=as.numeric(Subject)) %>%
 do(est.thalf(.,timevar="Time",depvar="conc",includeCmax="Y")) %>%
 ungroup()

# We need nominal time variable as well, so let's generate that.
ID <- as.numeric(Theoph$Subject)
NTAD <- c(0,0.3,0.5,1,2,4,5,7,9,12,24)
Theoph1 <- Theoph %>% mutate(NTAD=metrumrg::snap(Time, NTAD))

#let's say we want AUC0-8. We only have 7 and 9 hr concentrations, so we need to interpolate conc for 8 hr.
tc = Theoph1 %>%
group_by(Subject=as.numeric(Subject)) %>%
do(correct.time(.,nomtimevar="NTAD",timevar="Time",depvar="conc",
                 tau=,tstart=,tend=,teval=8,th=th,reg="sd"))

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"

#> Joining, by = "Subject"